requirements.txt
fastapi
uvicorn
pydantic
requests
joblib
numpy
pefile
python-whois
tldextract
xgboost

